Extracted details to separate component
